requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
openai>=1.12.0
langchain-openai>=0.0.5
langchain-community>=0.0.16
langchain-core>=0.1.17
chromadb>=0.4.22
langchain-chroma>=0.0.5
pymupdf>=1.23.8
